test(reportes): [T4.2.10] Probar filtros de fecha
Se valido en tiempo de ejecucion el correcto funcionamiento de las consultas SQL por dia, semana y mes, la actualizacion del TableView y el calculo de totales.

+++ src/org/esperanza/dao/ReporteVentasDao.java
@@ -35,7 +35,7 @@ public class ReporteVentasDao {
             while (rs.next()) {
                 Venta venta = new Venta();
                 venta.setIdVenta(rs.getInt("id_venta"));
-                venta.setTotal(rs.getDouble("total"));
+                venta.setTotal(rs.getBigDecimal("total"));
                 lista.add(venta);
             }
         }
